requirements.txt
django
django-allauth
django-allauth[socialaccount]
pymysql
google-generativeai==0.8.4
scikit-learn>=1.0.0
pandas>=1.3.0
joblib>=1.0.0
imbalanced-learn>=0.8.0
matplotlib>=3.0.0